tests/wpt/css-tests/cssom-1_dev/html/cssom-cssstyledeclaration-set.htm

<!DOCTYPE html>
<html><head>
        <title>CSSOM: CSSStyleDeclaration on HTMLElement represents inline style changes</title>
        <link href="mailto:paul.irish@gmail.com" rel="author" title="Paul Irish">
        <link href="http://www.w3.org/TR/cssom/#the-cssstyledeclaration-interface" rel="help">

        <link href="http://trac.webkit.org/export/120528/trunk/LayoutTests/fast/css/cssText-cache.html" rel="source">
        <meta content="dom" name="flags">

        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>

    <body>
        <div id="log"></div>

        <div id="box"></div>

        <script>

          var style = document.getElementById('box').style;

          test(function(){

            style.left = "10px";
            assert_equals(style.left, "10px", 'left property set on element\'s CSSStyleDeclaration Object');
            style.left = "20px";
            assert_equals(style.left, "20px", 'left property set on element\'s CSSStyleDeclaration Object');

          }, 'CSSStyleDeclaration on HTMLElement represents inline style changes');

        </script>


    

</body></html>
